Allow to distinguish between date/time representations
Date/time values can be represented as a number (epoch) or string for
transfer (iso8601). The openapi spec can now contain an integer field
with `format: date-time` (or `date`, resp.), which will be available
for custom type mappings. The types for scala and java are
unchanged (LocalDate/Time), but now one can change these based on what
was given in the specification. For this, the `TypeDef` class now
carries the `Type` it originated from. For Elm, the types are now
reflecting the base type (either `String` or a `Int`).